LayoutTests/imported/w3c/web-platform-tests/svg/animations/svglength-animation-invalid-value-9-expected.txt


FAIL Test SVGLength animation with invalid value: No spaces between number and unit in values attribute. assert_equals: expected 100 but got 300
